requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
diskcache==5.6.3
distro==1.9.0
docker==7.1.0
FLAML==2.5.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==1.26.4
openai==2.15.0
packaging==26.0
pyautogen==0.2.35
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
regex==2026.1.15
requests==2.32.5
sniffio==1.3.1
termcolor==3.3.0
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
